# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.0
paths:
...
  /api/v1/economy/wares:
    get:
      tags:
      - economy
      summary: List Ware Market
      description: Universe-wide ware supply/demand/price summary. Most under-supplied
        first.
      operationId: list_ware_market_api_v1_economy_wares_get
      parameters:
      - name: classification
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - shortage
            - balanced
            - surplus
            type: string
          - type: 'null'
          description: Filter to shortages, surpluses, or balanced wares
          title: Classification
        description: Filter to shortages, surpluses, or balanced wares
      - name: sort
        in: query
        required: false
        schema:
          enum:
          - net_demand
          - price_index
          - sell_qty
          - buy_qty
          type: string
          description: Sort key (descending)
          default: net_demand
          title: Sort
        description: Sort key (descending)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WareMarketRow'
                title: Response List Ware Market Api V1 Economy Wares Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/economy/wares/{ware_id}/stations:
    get:
      tags:
      - economy
      summary: Ware Offer Breakdown
      description: Per-station offers for one ware — where it's demanded vs supplied
        (and hoarding).
      operationId: ware_offer_breakdown_api_v1_economy_wares__ware_id__stations_get
      parameters:
      - name: ware_id
        in: path
        required: true
        schema:
          type: string
          title: Ware Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WareOfferRow'
                title: Response Ware Offer Breakdown Api V1 Economy Wares  Ware Id  Stations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/economy/accounts:
    get:
      tags:
      - economy
      summary: List Accounts
      description: Tracked money accounts with their latest net worth — feeds the
        net-worth picker.
      operationId: list_accounts_api_v1_economy_accounts_get
      parameters:
      - name: player_only
        in: query
        required: false
        schema:
          type: boolean
          description: Only accounts resolving to player assets
          default: false
          title: Player Only
        description: Only accounts resolving to player assets
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRow'
                title: Response List Accounts Api V1 Economy Accounts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/economy/networth:
    get:
      tags:
      - economy
      summary: Net Worth Series
      description: Net-worth-over-time samples for charting (the in-game player overview
        curve).
      operationId: net_worth_series_api_v1_economy_networth_get
      parameters:
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Pin one account; omit for all player accounts
          title: Owner
        description: Pin one account; omit for all player accounts
      - name: player_only
        in: query
        required: false
        schema:
          type: boolean
          description: Restrict to player accounts (ignored when owner set)
          default: true
          title: Player Only
        description: Restrict to player accounts (ignored when owner set)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NetWorthPointRow'
                title: Response Net Worth Series Api V1 Economy Networth Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/economy/pnl:
    get:
      tags:
      - economy
      summary: Ware Pnl
      description: Per-commodity profit/loss from the player's transactions, most
        profitable first.
      operationId: ware_pnl_api_v1_economy_pnl_get
      parameters:
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Scope to one or more ships/stations (buyer or seller)
          title: Owner
        description: Scope to one or more ships/stations (buyer or seller)
      - name: ware
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Scope to one or more ware_ids
          title: Ware
        description: Scope to one or more ware_ids
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          description: Only trades at or after this in-game time (seconds)
          title: Since
        description: Only trades at or after this in-game time (seconds)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WarePnlRow'
                title: Response Ware Pnl Api V1 Economy Pnl Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/economy/trades:
    get:
      tags:
      - economy
      summary: List Trades
      description: Transaction ledger, most recent first — where credits are earned
        and spent.
      operationId: list_trades_api_v1_economy_trades_get
      parameters:
      - name: ware
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Scope to one or more ware_ids
          title: Ware
        description: Scope to one or more ware_ids
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Trades where any of these ids is buyer or seller
          title: Owner
        description: Trades where any of these ids is buyer or seller
      - name: player_only
        in: query
        required: false
        schema:
          type: boolean
          description: Only trades a player asset is party to
          default: false
          title: Player Only
        description: Only trades a player asset is party to
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          description: Only trades at or after this in-game time (seconds)
          title: Since
        description: Only trades at or after this in-game time (seconds)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TradeRow'
                title: Response List Trades Api V1 Economy Trades Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccountRow:
      properties:
        owner:
          type: string
          title: Owner
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        kind:
          type: string
          title: Kind
        faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction
        is_player:
          type: boolean
          title: Is Player
        ship_role:
          anyOf:
          - type: string
          - type: 'null'
          title: Ship Role
        net_worth:
          anyOf:
          - type: integer
          - type: 'null'
          title: Net Worth
        net_worth_assets:
          anyOf:
          - type: integer
          - type: 'null'
          title: Net Worth Assets
        live_cash:
          anyOf:
          - type: integer
          - type: 'null'
          title: Live Cash
        account_amount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Account Amount
        account_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Account Min
        account_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Account Max
        latest_time:
          anyOf:
          - type: number
          - type: 'null'
          title: Latest Time
        event_count:
          type: integer
          title: Event Count
      additionalProperties: false
      type: object
      required:
      - owner
      - name
      - kind
      - faction
      - is_player
      - ship_role
      - net_worth
      - net_worth_assets
      - live_cash
      - account_amount
      - account_min
      - account_max
      - latest_time
      - event_count
      title: AccountRow
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NetWorthPointRow:
      properties:
        owner:
          type: string
          title: Owner
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        is_player:
          type: boolean
          title: Is Player
        time:
          type: number
          title: Time
        v:
          anyOf:
          - type: integer
          - type: 'null'
          title: V
        v2:
          anyOf:
          - type: integer
          - type: 'null'
          title: V2
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        delta:
          anyOf:
          - type: integer
          - type: 'null'
          title: Delta
        partner:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner
        partner_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Name
        partner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Faction
        partner_faction_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Faction Name
        partner_kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Kind
        partner_is_player:
          type: boolean
          title: Partner Is Player
      additionalProperties: false
      type: object
      required:
      - owner
      - name
      - is_player
      - time
      - v
      - v2
      - type
      - delta
      - partner
      - partner_name
      - partner_faction
      - partner_faction_name
      - partner_kind
      - partner_is_player
      title: NetWorthPointRow
...
    TradeRow:
      properties:
        time:
          type: number
          title: Time
        ware:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware
        ware_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware Name
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        price:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price
        quantity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quantity
        buyer:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer
        buyer_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Name
        buyer_is_player:
          type: boolean
          title: Buyer Is Player
        seller:
          anyOf:
          - type: string
          - type: 'null'
          title: Seller
        seller_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Seller Name
        seller_is_player:
          type: boolean
          title: Seller Is Player
      additionalProperties: false
      type: object
      required:
      - time
      - ware
      - ware_name
      - icon_url
      - price
      - quantity
      - buyer
      - buyer_name
      - buyer_is_player
      - seller
      - seller_name
      - seller_is_player
      title: TradeRow
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WareMarketRow:
      properties:
        ware_id:
          type: string
          title: Ware Id
        ware_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware Name
        group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
        ref_price:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ref Price
        sell_offers:
          type: integer
          title: Sell Offers
        sell_qty:
          type: integer
          title: Sell Qty
        best_buy_price:
          anyOf:
          - type: integer
          - type: 'null'
          title: Best Buy Price
        buy_offers:
          type: integer
          title: Buy Offers
        buy_qty:
          type: integer
          title: Buy Qty
        best_sell_price:
          anyOf:
          - type: integer
          - type: 'null'
          title: Best Sell Price
        avg_price:
          type: integer
          title: Avg Price
        net_demand:
          type: integer
          title: Net Demand
        price_index:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Index
        classification:
          type: string
          title: Classification
      additionalProperties: false
      type: object
      required:
      - ware_id
      - ware_name
      - group_id
      - ref_price
      - sell_offers
      - sell_qty
      - best_buy_price
      - buy_offers
      - buy_qty
      - best_sell_price
      - avg_price
      - net_demand
      - price_index
      - classification
      title: WareMarketRow
    WareOfferRow:
      properties:
        station_id:
          type: string
          title: Station Id
        station_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Station Name
        station_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Station Code
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        sector_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Id
        side:
          type: string
          title: Side
        price:
          type: integer
          title: Price
        quantity:
          type: integer
          title: Quantity
      additionalProperties: false
      type: object
      required:
      - station_id
      - station_name
      - station_code
      - owner_faction
      - sector_id
      - side
      - price
      - quantity
      title: WareOfferRow
    WarePnlRow:
      properties:
        ware:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware
        ware_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware Name
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        income:
          type: integer
          title: Income
        spend:
          type: integer
          title: Spend
        net:
          type: integer
          title: Net
        sell_count:
          type: integer
          title: Sell Count
        buy_count:
          type: integer
          title: Buy Count
        sell_qty:
          type: integer
          title: Sell Qty
        buy_qty:
          type: integer
          title: Buy Qty
      additionalProperties: false
      type: object
      required:
      - ware
      - ware_name
      - icon_url
      - income
      - spend
      - net
      - sell_count
      - buy_count
      - sell_qty
      - buy_qty
      title: WarePnlRow
...